pyproject.toml
[project]
name = "serverless-autoresearch"
version = "0.1.0"
description = "Parallel evolutionary autoresearch on SageMaker Spot Training"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "sagemaker>=2.232.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]
